chatbot/requirements.txt
langchain==0.0.316
openai
streamlit
pyautogen